+++ src/browserbehavior/DefaultBrowserBehavior.ts
@@ -6,6 +6,28 @@ import { detect } from 'detect-browser';
 export default class DefaultBrowserBehavior {
   private readonly browser = detect();
 
+  private browserSupport: { [id: string]: number } = {
+    chrome: 78,
+    'edge-chromium': 79,
+    electron: 7,
+    firefox: 60,
+    ios: 12,
+    safari: 12,
+    opera: 66,
+  };
+
+  private browserName: { [id: string]: string } = {
+    chrome: 'Google Chrome',
+    'edge-chromium': 'Microsoft Edge',
+    electron: 'Electron',
+    firefox: 'Mozilla Firefox',
+    ios: 'Safari iOS',
+    safari: 'Safari',
+    opera: 'Opera',
+  };
+
+  private chromeLike: string[] = ['chrome', 'edge-chromium', 'chromium-webview', 'opera'];
+
   version(): string {
     return this.browser.version;
   }
@@ -14,24 +36,25 @@ export default class DefaultBrowserBehavior {
     return parseInt(this.version().split('.')[0]);
   }
 
-  isEdgeChromium(): boolean {
-    return this.browser.name === 'edge-chromium';
-  }
-
-  isSafari(): boolean {
-    return this.browser.name === 'safari' || this.browser.name === 'ios';
-  }
-
-  isChrome(): boolean {
-    return this.browser.name === 'chrome';
-  }
-
-  isFirefox(): boolean {
-    return this.browser.name === 'firefox';
-  }
-
   name(): string {
     return this.browser.name;
+  }
+
+  hasChromiumWebRTC(): boolean {
+    for (const browser of this.chromeLike) {
+      if (browser === this.browser.name) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  hasFirefoxWebRTC(): boolean {
+    return this.isFirefox();
+  }
+
+  screenShareSendsOnlyKeyframes(): boolean {
+    return this.isFirefox();
   }
 
   requiresUnifiedPlan(): boolean {
@@ -39,7 +62,7 @@ export default class DefaultBrowserBehavior {
   }
 
   requiresIceCandidateGatheringTimeoutWorkaround(): boolean {
-    return this.isChrome();
+    return this.hasChromiumWebRTC();
   }
 
   requiresUnifiedPlanMunging(): boolean {
@@ -55,23 +78,36 @@ export default class DefaultBrowserBehavior {
   }
 
   requiresPromiseBasedWebRTCGetStats(): boolean {
-    return !this.isChrome();
+    return !this.hasChromiumWebRTC();
   }
 
-  isSupported(): boolean {
+  screenShareUnsupported(): boolean {
     if (this.isSafari()) {
-      return this.majorVersion() >= 13;
-    } else if (this.isChrome()) {
-      return this.majorVersion() >= 78;
-    } else if (this.isFirefox()) {
-      return this.majorVersion() >= 60;
-    } else if (this.isEdgeChromium()) {
-      return this.majorVersion() >= 79;
+      return true;
     }
     return false;
   }
 
+  isSupported(): boolean {
+    return this.majorVersion() >= this.browserSupport[this.browser.name];
+  }
+
   supportString(): string {
-    return 'Chromium 78+, Electron 7+, Microsoft Edge 79+, Google Chrome 78+, Mozilla Firefox 60+';
+    const s: string[] = [];
+    for (const k in this.browserSupport) {
+      s.push(`${this.browserName[k]} ${this.browserSupport[k]}+`);
+    }
+    return s.join(', ');
+  }
+
+  // These helpers should be kept private to encourage
+  // feature detection instead of browser detection.
+
+  private isSafari(): boolean {
+    return this.browser.name === 'safari' || this.browser.name === 'ios';
+  }
+
+  private isFirefox(): boolean {
+    return this.browser.name === 'firefox';
   }
 }